third_party/WebKit/LayoutTests/printing/print-media-recalc.html

<!DOCTYPE html>
<script src="../resources/js-test.js"></script>
<style>
@media print {
    body { font-size: 150px; orphans:1; widows:1; }
}
</style>
A<br>
B
<script>
// Check that the 150px font-size for body is applied for printing, causing
// room for only one line per page with a page height of 200px.

shouldBe("getComputedStyle(document.body, null).fontSize", "'16px'", true);

if (window.internals)
    shouldBe("internals.numberOfPages(800, 200)", "2");

shouldBe("getComputedStyle(document.body, null).fontSize", "'16px'");
</script>
