third_party/blink/web_tests/platform/generic/printing/print-media-recalc-expected.txt

PASS internals.numberOfPages(800, 200) is 2
PASS getComputedStyle(document.body, null).fontSize is '16px'
PASS successfullyParsed is true

TEST COMPLETE

A
B
